+++ view/SoftBodyView.java
@@ -11,7 +11,11 @@ import javax.swing.JFrame;
 import javax.swing.JPanel;
 
 import controller.SoftBodyController;
+import model.ModelConfig;
 import model.ReadOnlyModel;
+import model.SoftBodyModel;
+import model.masspoint.MassPoint;
+import model.masspoint.ReadOnlyMassPoint;
 import model.softbody.ReadOnlySoftBody;
 
 public class SoftBodyView extends JPanel implements Runnable, ViewConfig {
@@ -74,8 +78,43 @@ public class SoftBodyView extends JPanel implements Runnable, ViewConfig {
         g.setStroke(new BasicStroke(2));
 
         for (int i = 0; i < softBodies.size(); i++) {
-            softBodies.get(i).paintComponent(g);
+            List<ReadOnlyMassPoint> points = softBodies.get(i).getPoints();
+            ReadOnlySoftBody body = softBodies.get(i);
+
+            if (SoftBodyModel.fillSofbtody) {
+                g.setColor(Color.gray);
+                g.fillPolygon(body.getXArr(), body.getYArr(), points.size());
+            }
+
+            if (SoftBodyModel.drawSprings) {
+                g.setColor(Color.WHITE);
+                for (int j = 0; j < points.size(); j++) {
+                    g.drawLine((int) points.get(j).getPosition().getX(),
+                            (int) points.get(j).getPosition().getY(),
+                            (int) points.get((j + 1) % points.size()).getPosition().getX(),
+                            (int) points.get((j + 1) % points.size()).getPosition().getY());
+                }
+            }
+
+            if (SoftBodyModel.drawPoints) {
+                g.setColor(Color.red);
+
+                for (int j = 0; i < points.size(); i++) {
+
+                    g.fillOval((int) (points.get(i).getPositionX() - ModelConfig.POINT_SIZE),
+                            (int) (points.get(i).getPositionY() - ModelConfig.POINT_SIZE), (int) (2 *
+                                    ModelConfig.POINT_SIZE),
+                            (int) (2 * ModelConfig.POINT_SIZE));
+                }
+
+            }
+
+            // draw bounding box
+            if (SoftBodyModel.drawBoundingBox) {
+                g.setColor(Color.green);
+                g.drawRect(body.getBoundingBox().X(), body.getBoundingBox().Y(), body.getBoundingBox().width(),
+                        body.getBoundingBox().height());
+            }
         }
     }
-
 }